+++ src/pages/404.js
@@ -3,7 +3,7 @@ import NextLink from 'next/link';
 import { Box, Button, Container, Typography } from '@mui/material';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 
-const NotFound = () => (
+const Page = () => (
   <>
     <Head>
       <title>
@@ -73,4 +73,4 @@ const NotFound = () => (
   </>
 );
 
-export default NotFound;
+export default Page;