+++ src/pages/_app.js
@@ -1,11 +1,16 @@
+import { Fragment } from 'react';
 import Head from 'next/head';
 import { CacheProvider } from '@emotion/react';
-import LocalizationProvider from '@mui/lab/LocalizationProvider';
-import AdapterDateFns from '@mui/lab/AdapterDateFns';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { CssBaseline } from '@mui/material';
 import { ThemeProvider } from '@mui/material/styles';
+import { AuthConsumer, AuthProvider } from '../contexts/auth-context';
 import { createEmotionCache } from '../utils/create-emotion-cache';
+import { registerChartJs } from '../utils/register-chart-js';
 import { theme } from '../theme';
+
+registerChartJs();
 
 const clientSideEmotionCache = createEmotionCache();
 
@@ -28,7 +33,15 @@ const App = (props) => {
       <LocalizationProvider dateAdapter={AdapterDateFns}>
         <ThemeProvider theme={theme}>
           <CssBaseline />
-          {getLayout(<Component {...pageProps} />)}
+          <AuthProvider>
+            <AuthConsumer>
+              {
+                (auth) => auth.isLoading
+                  ? <Fragment />
+                  : getLayout(<Component {...pageProps} />)
+              }
+            </AuthConsumer>
+          </AuthProvider>
         </ThemeProvider>
       </LocalizationProvider>
     </CacheProvider>